@@ -7,15 +7,12 @@ module PactBroker
 
         property :method
         property :url
-        property :headers, getter: lambda { | _ | headers.empty? ? nil : self.redacted_headers }
+        property :headers, getter: lambda { | _ | self.redacted_headers.empty? ? nil : self.redacted_headers }
         property :body
         property :username
         property :password, getter: lambda { | _ | display_password }
 
 
-        def redacted_headers
-          represented.headers
-        end
       end
     end
   end
