@@ -34,7 +34,7 @@ module PactBroker
           if pacticipant
             @pacticipant = pacticipant_service.update params_with_string_keys.merge('name' => pacticipant_name)
           else
-            @pacticipant = pacticipant_service.create params_with_string_keys.merge('name' => pacticipant_name)
+            @pacticipant = pacticipant_service.create params.merge(:name => pacticipant_name)
             response.headers["Location"] = pacticipant_url(base_url, pacticipant)
           end
           response.body = to_json
