@@ -3,13 +3,18 @@ require 'pact_broker/configuration'
 require 'pact_broker/domain/verification'
 require 'pact_broker/api/contracts/verification_contract'
 require 'pact_broker/api/decorators/verification_decorator'
+require 'pact_broker/api/decorators/extended_verification_decorator'
 
 module PactBroker
   module Api
     module Resources
       class Verification < BaseResource
         def content_types_provided
-          [["application/hal+json", :to_json], ["application/json", :to_json]]
+          [
+            ["application/hal+json", :to_json],
+            ["application/json", :to_json],
+            ["application/vnd.pactbrokerextended.v1+json", :to_extended_json]
+          ]
         end
 
         # Remember to update latest_verification_id_for_pact_version_and_provider_version
@@ -26,6 +31,10 @@ module PactBroker
           decorator_for(verification).to_json(user_options: {base_url: base_url})
         end
 
+        def to_extended_json
+          extended_decorator_for(verification).to_json(user_options: {base_url: base_url})
+        end
+
         private
 
         def verification
@@ -34,6 +43,10 @@ module PactBroker
 
         def decorator_for model
           PactBroker::Api::Decorators::VerificationDecorator.new(model)
+        end
+
+        def extended_decorator_for model
+          PactBroker::Api::Decorators::ExtendedVerificationDecorator.new(model)
         end
       end
     end
