@@ -1,26 +1,47 @@
 require 'pact_broker/api/resources/base_resource'
 require 'pact_broker/services'
 require 'pact_broker/api/decorators/webhook_execution_result_decorator'
+require 'pact_broker/api/resources/webhook_resource_methods'
 require 'pact_broker/constants'
+require 'pact_broker/webhooks/execution_configuration'
+require 'pact_broker/api/resources/webhook_execution_methods'
 
 module PactBroker
   module Api
     module Resources
       class WebhookExecution < BaseResource
+        include WebhookResourceMethods
+        include WebhookExecutionMethods
+
+        def content_types_accepted
+          [["application/json"]]
+        end
+
+        def content_types_provided
+          [["application/hal+json"]]
+        end
 
         def allowed_methods
           ["POST", "OPTIONS"]
         end
 
         def process_post
-          webhook_execution_result = webhook_service.test_execution(webhook, webhook_options)
+          webhook_execution_result = webhook_service.test_execution(webhook, webhook_execution_configuration)
           response.headers['Content-Type'] = 'application/hal+json;charset=utf-8'
-          response.body = post_response_body webhook_execution_result
+          response.body = post_response_body(webhook_execution_result)
           true
         end
 
         def resource_exists?
           webhook
+        end
+
+        def malformed_request?
+          if uuid
+            false
+          else
+            webhook_validation_errors?(webhook)
+          end
         end
 
         private
@@ -30,7 +51,13 @@ module PactBroker
         end
 
         def webhook
-          @webhook ||= webhook_service.find_by_uuid uuid
+          @webhook ||= begin
+            if uuid
+              webhook_service.find_by_uuid uuid
+            else
+              build_unsaved_webhook
+            end
+          end
         end
 
         def uuid
@@ -38,22 +65,14 @@ module PactBroker
         end
 
         def user_options
-          {
-            base_url: base_url,
+          decorator_context(
             webhook: webhook,
             show_response: PactBroker.configuration.show_webhook_response?
-          }
+          )
         end
 
-        def webhook_options
-          {
-            logging_options: {
-              show_response: PactBroker.configuration.show_webhook_response?
-            },
-            webhook_context: {
-              base_url: base_url
-            }
-          }
+        def build_unsaved_webhook
+          Decorators::WebhookDecorator.new(PactBroker::Domain::Webhook.new).from_json(request_body)
         end
       end
     end
