@@ -11,7 +11,7 @@ module PactBroker
     class Pact
 
       # The ID is the pact_publication ID
-      attr_accessor :id, :provider, :consumer_version, :consumer, :created_at, :json_content, :consumer_version_number, :revision_number, :pact_version_sha, :latest_verification
+      attr_accessor :id, :provider, :consumer_version, :consumer, :created_at, :json_content, :consumer_version_number, :revision_number, :pact_version_sha, :latest_verification, :head_tag_names
       def initialize attributes
         attributes.each_pair do | key, value |
           self.send(key.to_s + "=", value)
