@@ -1,9 +1,10 @@
-require 'pact_broker/db'
-require 'pact_broker/repositories/helpers'
 require 'json'
+require 'sequel'
+require 'pact_broker/repositories/helpers'
+require 'pact_broker/tags/tag_with_latest_flag'
+
 
 module PactBroker
-
   module Domain
     class Verification < Sequel::Model
 
@@ -12,6 +13,7 @@ module PactBroker
       associate(:many_to_one, :provider_version, class: "PactBroker::Domain::Version", key: :provider_version_id, primary_key: :id)
       associate(:many_to_one, :provider, class: "PactBroker::Domain::Pacticipant", key: :provider_id, primary_key: :id)
       associate(:many_to_one, :consumer, class: "PactBroker::Domain::Pacticipant", key: :consumer_id, primary_key: :id)
+      associate(:one_to_many, :provider_version_tags, :class => "PactBroker::Tags::TagWithLatestFlag", primary_key: :provider_version_id, key: :version_id)
       plugin :serialization, :json, :test_results
 
       def before_create
