@@ -54,29 +54,13 @@ module PactBroker
 
       def execute pact, verification, options
         logger.info "Executing #{self}"
-        webhook_request = request.build(pact: pact, verification: verification, webhook_context: options.fetch(:webhook_context))
-        http_response = nil
-        error = nil
-        begin
-          http_response = webhook_request.execute
-        rescue StandardError => e
-          error = e
-        end
+        webhook_request = request.build(template_parameters(pact, verification, options))
+        http_response, error = execute_request(webhook_request)
 
         PactBroker::Webhooks::WebhookExecutionResult.new(
           webhook_request.http_request,
           http_response,
           generate_logs(webhook_request, http_response, error, options.fetch(:logging_options)),
-          error
-        )
-      end
-
-      def generate_logs(webhook_request, http_response, error, logging_options)
-        webhook_request_logger = PactBroker::Webhooks::WebhookRequestLogger.new(logging_options)
-        webhook_request_logger.log(
-          uuid,
-          webhook_request,
-          http_response,
           error
         )
       end
@@ -99,6 +83,33 @@ module PactBroker
 
       def trigger_on_provider_verification_published?
         events.any?(&:provider_verification_published?)
+      end
+
+      private
+
+      def execute_request(webhook_request)
+        http_response = nil
+        error = nil
+        begin
+          http_response = webhook_request.execute
+        rescue StandardError => e
+          error = e
+        end
+        return http_response, error
+      end
+
+      def template_parameters(pact, verification, options)
+        PactBroker::Webhooks::PactAndVerificationParameters.new(pact, verification, options.fetch(:webhook_context)).to_hash
+      end
+
+      def generate_logs(webhook_request, http_response, error, logging_options)
+        webhook_request_logger = PactBroker::Webhooks::WebhookRequestLogger.new(logging_options)
+        webhook_request_logger.log(
+          uuid,
+          webhook_request,
+          http_response,
+          error
+        )
       end
     end
   end
