@@ -23,11 +23,13 @@ module PactBroker
         )
       end
 
-      def self.for_pacticipant_and_version(pacticipant, version, original_selector, type)
+      def self.for_pacticipant_and_version(pacticipant, version, pact_publication_ids = [], verification_ids = [], original_selector, type)
         ResolvedSelector.new(
           pacticipant_id: pacticipant.id,
           pacticipant_name: pacticipant.name,
           pacticipant_version_id: version.id,
+          pact_publication_ids: pact_publication_ids,
+          verification_ids: verification_ids,
           pacticipant_version_number: version.number,
           latest: original_selector[:latest],
           tag: original_selector[:tag],
