@@ -31,14 +31,26 @@ module PactBroker
         Content.from_hash(SortContent.call(pact_hash))
       end
 
-      def with_ids
+      def with_test_results(test_results)
         new_pact_hash = pact_hash.dup
         if interactions && interactions.is_a?(Array)
-          new_pact_hash['interactions'] = add_ids(interactions)
+          new_pact_hash['interactions'] = merge_verification_results(interactions, test_results)
         end
 
         if messages && messages.is_a?(Array)
-          new_pact_hash['messages'] = add_ids(messages)
+          new_pact_hash['messages'] = merge_verification_results(messages, test_results)
+        end
+        Content.from_hash(new_pact_hash)
+      end
+
+      def with_ids(overwrite_existing_id = true)
+        new_pact_hash = pact_hash.dup
+        if interactions && interactions.is_a?(Array)
+          new_pact_hash['interactions'] = add_ids(interactions, overwrite_existing_id)
+        end
+
+        if messages && messages.is_a?(Array)
+          new_pact_hash['messages'] = add_ids(messages, overwrite_existing_id)
         end
         Content.from_hash(new_pact_hash)
       end
@@ -64,6 +76,10 @@ module PactBroker
         pact_hash.is_a?(Hash) ? pact_hash['interactions'] : nil
       end
 
+      def messages_or_interactions
+        messages || interactions
+      end
+
       def pact_specification_version
         maybe_pact_specification_version_1 = pact_hash['metadata']['pactSpecification']['version'] rescue nil
         maybe_pact_specification_version_2 = pact_hash['metadata']['pact-specification']['version'] rescue nil
@@ -75,13 +91,17 @@ module PactBroker
 
       attr_reader :pact_hash
 
-      def add_ids(interactions)
+      def add_ids(interactions, overwrite_existing_id)
         interactions.map do | interaction |
           if interaction.is_a?(Hash)
-            # just in case there is a previous ID in there
-            interaction_without_id = interaction.reject { |k, _| k == "_id" }
-            # make the _id the first key in the hash when rendered to JSON
-            { "_id" => generate_interaction_sha(interaction_without_id) }.merge(interaction)
+            if !interaction.key?("_id") || overwrite_existing_id
+              # just in case there is a previous ID in there
+              interaction_without_id = interaction.reject { |k, _| k == "_id" }
+              # make the _id the first key in the hash when rendered to JSON
+              { "_id" => generate_interaction_sha(interaction_without_id) }.merge(interaction)
+            else
+              interaction
+            end
           else
             interaction
           end
