@@ -164,10 +164,9 @@ module PactBroker
       end
 
       def merge_consumer_version_info(webhook_options, pact)
-        webhook_context = webhook_options.fetch(:webhook_context, {}).merge(
-          consumer_version_tags: pact.consumer_version_tag_names
-        )
-        webhook_options.merge(webhook_context: webhook_context)
+        execution_configuration = webhook_options[:webhook_execution_configuration]
+                                    .with_webhook_context(consumer_version_tags: pact.consumer_version_tag_names)
+        webhook_options.merge(webhook_execution_configuration: execution_configuration)
       end
     end
   end
