@@ -1,3 +1,3 @@
 module PactBroker
-  VERSION = '2.32.0'
+  VERSION = '2.38.1'
 end
