@@ -14,7 +14,7 @@ module PactBroker
         if pact_is_new_or_newly_tagged_or_pact_has_changed_since_previous_version?(pact)
           webhook_service.trigger_webhooks pact, nil, PactBroker::Webhooks::WebhookEvent::CONTRACT_CONTENT_CHANGED, webhook_options
         else
-          logger.debug "Pact content has not changed since previous version, not triggering webhooks for changed content"
+          logger.info "Pact content has not changed since previous version, not triggering webhooks for changed content"
         end
       end
 
@@ -22,10 +22,10 @@ module PactBroker
         webhook_service.trigger_webhooks updated_pact, nil, PactBroker::Webhooks::WebhookEvent::CONTRACT_PUBLISHED, webhook_options
         # TODO this should use the sha!
         if existing_pact.pact_version_sha != updated_pact.pact_version_sha
-          logger.debug "Existing pact for version #{existing_pact.consumer_version_number} has been updated with new content, triggering webhooks for changed content"
+          logger.info "Existing pact for version #{existing_pact.consumer_version_number} has been updated with new content, triggering webhooks for changed content"
           webhook_service.trigger_webhooks updated_pact, nil, PactBroker::Webhooks::WebhookEvent::CONTRACT_CONTENT_CHANGED, webhook_options
         else
-          logger.debug "Pact content has not changed since previous revision, not triggering webhooks for changed content"
+          logger.info "Pact content has not changed since previous revision, not triggering webhooks for changed content"
         end
       end
 
@@ -60,7 +60,7 @@ module PactBroker
               end
             end
           end
-          logger.debug("Webhook triggered for the following reasons: #{messages.join(',')}" )
+          logger.info("Webhook triggered for the following reasons: #{messages.join(',')}" )
         end
       end
     end
