@@ -2,6 +2,7 @@ require 'pact_broker/domain/webhook_request_header'
 require 'pact_broker/webhooks/render'
 require 'cgi'
 require 'pact_broker/domain/webhook_request'
+require 'pact_broker/string_refinements'
 
 module PactBroker
   module Webhooks
@@ -9,6 +10,8 @@ module PactBroker
 
       include PactBroker::Logging
       include PactBroker::Messages
+      using PactBroker::StringRefinements
+
       HEADERS_TO_REDACT = [/authorization/i, /token/i]
 
       attr_accessor :method, :url, :headers, :body, :username, :password, :uuid
@@ -27,40 +30,30 @@ module PactBroker
         @uuid = attributes[:uuid]
       end
 
-      def build(context)
-        template_params = PactBroker::Webhooks::PactAndVerificationParameters.new(context[:pact], context[:verification], context[:webhook_context]).to_hash
+      def build(template_params)
         attributes = {
           method: http_method,
           url: build_url(template_params),
-          headers: headers,
-          username: username,
-          password: password,
+          headers: build_headers(template_params),
+          username: build_string(username, template_params),
+          password: build_string(password, template_params),
           uuid: uuid,
           body: build_body(template_params)
         }
         PactBroker::Domain::WebhookRequest.new(attributes)
       end
 
-      def build_url(template_params)
-        URI(PactBroker::Webhooks::Render.call(url, template_params){ | value | CGI::escape(value) if !value.nil? } ).to_s
-      end
-
-      def build_body(template_params)
-        body_string = String === body ? body : body.to_json
-        PactBroker::Webhooks::Render.call(body_string, template_params)
-      end
-
       def description
-        "#{http_method.upcase} #{URI(url.gsub(PactBroker::Webhooks::Render::TEMPLATE_PARAMETER_REGEXP, 'placeholder')).host}"
+        "#{http_method.upcase} #{URI(PactBroker::Webhooks::Render.render_with_placeholder(url)).host}"
       end
 
       def display_password
-        password.nil? ? nil : "**********"
+        password.nil? ? nil : (PactBroker::Webhooks::Render.includes_parameter?(password) ? password : "**********")
       end
 
       def redacted_headers
         headers.each_with_object({}) do | (name, value), new_headers |
-          redact = HEADERS_TO_REDACT.any?{ | pattern | name =~ pattern }
+          redact = HEADERS_TO_REDACT.any?{ | pattern | name =~ pattern }  && !PactBroker::Webhooks::Render.includes_parameter?(value)
           new_headers[name] = redact ? "**********" : value
         end
       end
@@ -69,10 +62,31 @@ module PactBroker
         @headers = Rack::Utils::HeaderHash.new(headers)
       end
 
-      private
 
       def to_s
         "#{method.upcase} #{url}, username=#{username}, password=#{display_password}, headers=#{redacted_headers}, body=#{body}"
+      end
+
+      private
+
+      def build_url(template_params)
+        URI(PactBroker::Webhooks::Render.call(url, template_params){ | value | CGI::escape(value) if !value.nil? } ).to_s
+      end
+
+      def build_body(template_params)
+        body_string = String === body ? body : body.to_json
+        build_string(body_string, template_params)
+      end
+
+      def build_headers(template_params)
+        headers.each_with_object(Rack::Utils::HeaderHash.new) do | (key, value), new_headers |
+          new_headers[key] = build_string(value, template_params)
+        end
+      end
+
+      def build_string(string, template_params)
+        return string if string.nil? || string.blank?
+        PactBroker::Webhooks::Render.call(string, template_params)
       end
     end
   end
