@@ -4,11 +4,8 @@ require 'pact_broker/api/resources/webhook_execution'
 module PactBroker
   module Api
     module Resources
-
       describe WebhookExecution do
-
         describe "POST" do
-
           before do
             allow(PactBroker::Webhooks::Service).to receive(:find_by_uuid).and_return(webhook)
           end
@@ -33,28 +30,29 @@ module PactBroker
             let(:pact) { instance_double("PactBroker::Domain::Pact") }
             let(:consumer_name) { "foo" }
             let(:provider_name) { "bar" }
-            let(:webhook_options) do
-              {
-                logging_options: {
-                  show_response: false,
-                },
-                webhook_context: {
-                  base_url: "http://example.org"
-                }
-              }
-            end
+            let(:webhook_execution_configuration) { instance_double(PactBroker::Webhooks::ExecutionConfiguration) }
 
             before do
               allow(PactBroker::Webhooks::Service).to receive(:test_execution).and_return(execution_result)
               allow(PactBroker::Api::Decorators::WebhookExecutionResultDecorator).to receive(:new).and_return(decorator)
+              allow_any_instance_of(WebhookExecution).to receive(:webhook_execution_configuration).and_return(webhook_execution_configuration)
             end
 
             it "executes the webhook" do
-              expect(PactBroker::Webhooks::Service).to receive(:test_execution).with(webhook, webhook_options)
+              expect(PactBroker::Webhooks::Service).to receive(:test_execution).with(webhook, webhook_execution_configuration)
               subject
             end
 
             context "when execution is successful" do
+              let(:expected_user_options) do
+                {
+                  resource_url: 'http://example.org/webhooks/some-uuid/execute',
+                  base_url: 'http://example.org',
+                  webhook: webhook,
+                  show_response: 'foo',
+                }
+              end
+
               it "returns a 200 JSON response" do
                 subject
                 expect(last_response).to be_a_hal_json_success_response
@@ -62,7 +60,7 @@ module PactBroker
 
               it "generates a JSON response body for the execution result" do
                 allow(PactBroker.configuration).to receive(:show_webhook_response?).and_return('foo')
-                expect(decorator).to receive(:to_json).with(user_options: { base_url: 'http://example.org', webhook: webhook, show_response: 'foo' })
+                expect(decorator).to receive(:to_json).with(user_options: expected_user_options)
                 subject
               end
 
