@@ -4,6 +4,8 @@ module PactBroker
   module Webhooks
     describe RedactLogs do
       describe ".call" do
+        let(:values) { [] }
+
         let(:string) do
           "Authorization: foo\nX-Thing: bar"
         end
@@ -25,23 +27,32 @@ module PactBroker
         end
 
         it "hides the value of the Authorization header" do
-          expect(RedactLogs.call(string)).to eq "Authorization: [REDACTED]\nX-Thing: bar"
+          expect(RedactLogs.call(string, values)).to eq "Authorization: [REDACTED]\nX-Thing: bar"
         end
 
         it "hides the value of the X-Authorization header" do
-          expect(RedactLogs.call(x_auth_string)).to eq "X-Authorization: [REDACTED]\nX-Thing: bar"
+          expect(RedactLogs.call(x_auth_string, values)).to eq "X-Authorization: [REDACTED]\nX-Thing: bar"
         end
 
         it "hides the value of the X-Auth-Token header" do
-          expect(RedactLogs.call(x_auth_token)).to eq "X-Auth-Token: [REDACTED]\nX-Thing: bar"
+          expect(RedactLogs.call(x_auth_token, values)).to eq "X-Auth-Token: [REDACTED]\nX-Thing: bar"
         end
 
         it "hides the value of the X-Authorization-Token header" do
-          expect(RedactLogs.call(x_authorization_token)).to eq "X-Authorization-Token: [REDACTED]\nX-Thing: bar"
+          expect(RedactLogs.call(x_authorization_token, values)).to eq "X-Authorization-Token: [REDACTED]\nX-Thing: bar"
         end
 
         it "hides the value of the authorization header" do
-          expect(RedactLogs.call(string_lower)).to eq "authorization: [REDACTED]\nX-Thing: bar"
+          expect(RedactLogs.call(string_lower, values)).to eq "authorization: [REDACTED]\nX-Thing: bar"
+        end
+
+        context "with values" do
+          let(:values) { %w[foo bar] }
+          let(:string) { "blahfoo\nbar wiffle" }
+
+          it "hides the passed in values" do
+            expect(RedactLogs.call(string, values)).to eq "blah********\n******** wiffle"
+          end
         end
       end
     end
